move usmqeconfig into usmqe module
the code here is no longer a pytest plugin
