+++ usmqe/api/etcdapi/etcdapi.py
@@ -8,7 +8,7 @@ import time
 import requests
 import pytest
 from usmqe.api.base import ApiBase
-from plugin.usmqe_config import UsmConfig
+from usmqe.usmqeconfig import UsmConfig
 
 LOGGER = pytest.get_logger("etcdapi", module=True)
 config = UsmConfig()
