+++ usmqe/api/grafanaapi/grafanaapi.py
@@ -8,7 +8,7 @@ import requests
 import pytest
 from difflib import Differ
 from usmqe.api.base import ApiBase
-from plugin.usmqe_config import UsmConfig
+from usmqe.usmqeconfig import UsmConfig
 
 LOGGER = pytest.get_logger("grafanaapi", module=True)
 config = UsmConfig()
