+++ usmqe/api/tendrlapi/user.py
@@ -6,7 +6,7 @@ import json
 import requests
 import pytest
 from usmqe.api.tendrlapi.common import TendrlApi
-from plugin.usmqe_config import UsmConfig
+from usmqe.usmqeconfig import UsmConfig
 
 LOGGER = pytest.get_logger("tendrlapi.user", module=True)
 USERDATA_KEYS = {'email', 'name', 'role', 'username', 'email_notifications'}
