+++ usmqe_tests/api/conftest.py
@@ -5,7 +5,7 @@ import time
 import pytest
 
 from usmqe.api.tendrlapi.common import TendrlAuth, login, logout, TendrlApi
-from plugin.usmqe_config import UsmConfig
+from usmqe.usmqeconfig import UsmConfig
 
 config = UsmConfig()
